Added bug back in

+++ Project_Archive/ProjectArchive.WPF/MainWindow.xaml.cs
@@ -225,8 +225,8 @@ public class Game
         for (var i = 0; i < numIterations; i++)
         {
             //gravity should lean toward negative
-            //var delta = random.Next(-10, 5)/100;
-            var delta = (double)random.Next(-10, 5) / 100;
+            var delta = random.Next(-10, 5)/100;
+            //var delta = (double)random.Next(-10, 5) / 100;
             endingOutput += delta;
         }
         if (endingOutput >= userInputGuess)
